requirements.txt
# Only keep development requirements here.  Actual dependencies belong
# in setup.py.
#
# Maintain manulally, avoid `pip freeze > requirements.py`.
-e .
Django==1.7.1
MySQL-python==1.2.5
amqp==1.4.6
anyjson==0.3.3
astroid==1.3.2
billiard==3.3.0.19
boto==2.34.0
celery==3.1.17
djangorestframework==2.4.4
drf-compound-fields==0.2.2
elasticsearch==1.2.0
flake8==2.2.5
httplib2==0.9
kombu==3.0.24
logilab-common==0.63.0
mccabe==0.2.1
mock==1.0.1
nexmomessage==0.1.1
pep257==0.3.2
pep8==1.5.7
py==1.4.26
py3kwarn==0.4.4
pyflakes==0.8.1
pylint==1.4.0
pylint-django==0.5.5
pylint-plugin-utils==0.2.2
pyrabbit==1.1.0
pytest==2.6.4
pytest-mock==0.4.0
pytz==2014.9
requests==2.4.3
six==1.8.0
tox==1.8.1
virtualenv==1.11.6
wsgiref==0.1.2
